Move and expand document chain test
