+++ identity-iota/src/chain/diff_chain.rs
@@ -208,120 +208,3 @@ impl From<DiffChain> for Vec<DiffMessage> {
     diff_chain.inner
   }
 }
-
-#[cfg(test)]
-mod test {
-  use identity_core::common::Timestamp;
-  use identity_core::crypto::KeyPair;
-  use identity_did::did::CoreDIDUrl;
-  use identity_did::did::DID;
-  use identity_did::verification::MethodBuilder;
-  use identity_did::verification::MethodData;
-  use identity_did::verification::MethodRef;
-  use identity_did::verification::MethodType;
-
-  use crate::chain::DocumentChain;
-  use crate::chain::IntegrationChain;
-  use crate::diff::DiffMessage;
-  use crate::document::IotaDocument;
-  use crate::document::ResolvedIotaDocument;
-  use crate::tangle::MessageId;
-  use crate::tangle::TangleRef;
-
-  #[test]
-  fn test_diff_chain() {
-    let mut chain: DocumentChain;
-    let mut keys: Vec<KeyPair> = Vec::new();
-
-    // =========================================================================
-    // Create Initial Document
-    // =========================================================================
-    {
-      let keypair: KeyPair = KeyPair::new_ed25519().unwrap();
-      let mut document: IotaDocument = IotaDocument::new(&keypair).unwrap();
-      document
-        .sign_self(keypair.private(), &document.default_signing_method().unwrap().id())
-        .unwrap();
-      let mut resolved: ResolvedIotaDocument = ResolvedIotaDocument::from(document);
-      resolved.set_message_id(MessageId::new([8; 32]));
-      chain = DocumentChain::new(IntegrationChain::new(resolved).unwrap());
-      keys.push(keypair);
-    }
-
-    assert_eq!(
-      chain.current().document.metadata.proof().unwrap().verification_method(),
-      format!("#{}", IotaDocument::DEFAULT_METHOD_FRAGMENT)
-    );
-
-    // =========================================================================
-    // Push Integration Chain Update
-    // =========================================================================
-    {
-      let mut new: ResolvedIotaDocument = chain.current().clone();
-      let keypair: KeyPair = KeyPair::new_ed25519().unwrap();
-
-      // Replace the capability invocation signing key (one step key rotation).
-      let signing_method: MethodRef = MethodBuilder::default()
-        .id(CoreDIDUrl::from(chain.id().to_url().join("#key-2").unwrap()))
-        .controller(chain.id().clone().into())
-        .key_type(MethodType::Ed25519VerificationKey2018)
-        .key_data(MethodData::new_multibase(keypair.public()))
-        .build()
-        .map(Into::into)
-        .unwrap();
-
-      unsafe {
-        new.document.core_document_mut().capability_invocation_mut().clear();
-        new
-          .document
-          .core_document_mut()
-          .capability_invocation_mut()
-          .append(signing_method);
-      }
-
-      new.document.metadata.updated = Timestamp::now_utc();
-      new.document.metadata.previous_message_id = *chain.integration_message_id();
-
-      // Sign the update using the old document.
-      assert!(chain
-        .current()
-        .document
-        .sign_data(
-          &mut new.document,
-          keys[0].private(),
-          chain.current().document.default_signing_method().unwrap().id(),
-        )
-        .is_ok());
-      assert_eq!(
-        chain.current().document.metadata.proof().unwrap().verification_method(),
-        format!("#{}", IotaDocument::DEFAULT_METHOD_FRAGMENT)
-      );
-
-      keys.push(keypair);
-      assert!(chain.try_push_integration(new).is_ok());
-    }
-
-    // =========================================================================
-    // Push Diff Chain Update
-    // =========================================================================
-    {
-      let new: ResolvedIotaDocument = {
-        let mut this: ResolvedIotaDocument = chain.current().clone();
-        this.document.properties_mut().insert("foo".into(), 123.into());
-        this.document.properties_mut().insert("bar".into(), 456.into());
-        this.document.metadata.updated = Timestamp::now_utc();
-        this
-      };
-
-      // Sign using the new key added in the previous integration chain update.
-      let message_id = *chain.diff_message_id();
-      let mut diff: DiffMessage = chain
-        .current()
-        .document
-        .diff(&new.document, message_id, keys[1].private(), "#key-2")
-        .unwrap();
-      diff.set_message_id(message_id);
-      assert!(chain.try_push_diff(diff).is_ok());
-    }
-  }
-}
